Move private 3D arrow functionality to it's own module (#766)
